pyproject.toml
[tool.poetry]
name = "wallet"
version = "0.1.0"
description = ""
authors = ["chedug <nuralyhug@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.7"
pre-commit = "^3.3.3"
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
interrogate = "^1.5.0"
djangorestframework-simplejwt = "^5.2.2"
httpie = "^3.2.2"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
target-version = ['py310']
include = '\.pyi?$'
line-length = 79
